File Handling

File Handling

File Handling

Functions

Index

Functions

ECHOAnalysis.add_batch_info — Method.

add_batch_info(folder::AbstractString)

Add batch info to merged/ folders.

Example: batch001/metaplan2/merged/metaphlan2_taxonomic_profiles.tsv to batch001/metaplan2/merged/batch001_metaphlan2_taxonomic_profiles.tsv

source

ECHOAnalysis.merge_tables — Method.

merge_tables(folder, dataroot, fltr; suffix=fltr)

Recurssively walk through a folder and merge tables inside folders named dataroot with a given filter fltr. The fltr is assumed to be the end of the filename and removed, but a different suffix may be provided.

Usage: mergetables("biobakery", "metaphlan2", "profile.tsv")

source

Index

ECHOAnalysis.add_batch_info

ECHOAnalysis.breastfeeding

ECHOAnalysis.formulafeeding

ECHOAnalysis.getfocusmetadata

ECHOAnalysis.getfocusmetadata

ECHOAnalysis.getmetadata

ECHOAnalysis.getmetadatum

ECHOAnalysis.merge_tables

ECHOAnalysis.metacolor

ECHOAnalysis.numberify

ECHOAnalysis.resolve_letter_timepoint

ECHOAnalysis.resolve_sampleID

ECHOAnalysis.samplelessthan

ECHOAnalysis.uniquesamples