Timepoints and Samples

Timepoints and Samples

Functions

Index

Functions

ECHOAnalysis.AbstractTimepoint — Type

Represents a given timepoint data collection or sample.

Required fields:

id – String: the unique identifier

subject – Int: the subject ID

timepoint – Int: The timepoint. For some data types, this will be 0 indicating that it applies to all timepoints for that subject.

ECHOAnalysis.StoolSample — Type

Represents a given stool sample

Fields:

id – String: the unique sample identifier

subject – Int: the subject ID

timepoint – Int: The timepoint. For some metadata, this will be 0

replicate – String: replicate / aliquot ID (eg "2B")

type – String: "omnigene" or "ethanol"

ECHOAnalysis.iskid — Method

iskid(s::Union{AbstractString, StoolSample})

Return true if StoolSample (or a String that can be converted to StoolSample) refers to a child sample, false otherwise.

ECHOAnalysis.ismom — Method

ismom(s::Union{AbstractString, StoolSample})

Return true if StoolSample (or a String that can be converted to StoolSample) refers to a maternal sample, false otherwise.

ECHOAnalysis.isstoolsample — Method

isstoolsample(sid::Union{AbstractString, AbstractTimepoint, StoolSample})

Returns true if sid is a StoolSample, or if an AbstractTimepoint's id field or AbstractString conforms to the expected form of an ECHO RESONANCE stool sample.

ECHOAnalysis.replicateid — Method

replicateid(s::Union{AbstractString, StoolSample})

Get the replicate field from a StoolSample (or a String that can be converted to StoolSample).

ECHOAnalysis.resolve_letter_timepoint — Method

resolve_letter_timepoint(sid::AbstractString)
resolve_letter_timepoint(sid::Missing)

Deals with sampleIDs in which the timepoint is given as a letter. Eg. 40b indicates timepoint 2 from subject ID 40. If no letter is provided, timepoint is assumed to be 1.

ECHOAnalysis.sampleid — Method

sampleid(s::AbstractTimepoint)

Get the id field from an AbstractTimepoint.

ECHOAnalysis.sampletype — Method

sampletype(s::Union{AbstractString, StoolSample})

Get the type field from a StoolSample (or a String that can be converted to StoolSample).

ECHOAnalysis.stoolsample — Method

stoolsample(sid::Union{AbstractString,Symbol})

Parse a sample ID and create a StoolSample.

Example: For the sample C0040_3F_1A:

C0040 means ""Child, subject ID 40"

3F means "timepoint 3", Fecal (omnigene). May also be E for Ethanol.

1A means "replicate 1, aliquot A"

Hyphens in sample names are converted to underscores (so "C00011F1A" == "C0001-1F-1A")

Samples may also end with _S\d+ (well #s from sequencing) which are ignored

ECHOAnalysis.subject — Method

subject(s::AbstractTimepoint)

Get the subject field from an AbstractTimepoint.

ECHOAnalysis.timepoint — Method

timepoint(s::AbstractTimepoint)

Get the timepoint field from an AbstractTimepoint.

Index

ECHOAnalysis.AbstractTimepoint

ECHOAnalysis.StoolSample

ECHOAnalysis.add_functional_profiles

ECHOAnalysis.add_taxonomic_profiles

ECHOAnalysis.breastfeeding

ECHOAnalysis.iskid

ECHOAnalysis.ismom

ECHOAnalysis.isstoolsample

ECHOAnalysis.numberify

ECHOAnalysis.replicateid

ECHOAnalysis.resolve_letter_timepoint

ECHOAnalysis.sampleid

ECHOAnalysis.sampletable

ECHOAnalysis.sampletype

ECHOAnalysis.sqlprofile

ECHOAnalysis.stoolsample

ECHOAnalysis.subject

ECHOAnalysis.timepoint

ECHOAnalysis.uniquetimepoints

ECHOAnalysis.widemetadata