Dealing with Metadata

Subject and Sample Metadata

Subject and Sample Metadata

Functions

Index

Functions

ECHOAnalysis.breastfeeding — Method.

breastfeeding(row::DataFrameRow)

Checks a wide-form metadata row for breastfeeding information. Returns true if any of the following are true:

:milkFeedingMethods contains "breast"

:exclusivelyNursed is "Yes" (or "yes")

Both :exclusivelyNursed and :exclusiveFormulaFed are "No" (or "no")

Any of the following have values > 0:

:typicalNumberOfFeedsFromBreast

:typicalNumberOfEpressedMilkFeeds

:lengthExclusivelyNursedMonths

:noLongerFeedBreastmilkAge

Otherwise returns false.

source

ECHOAnalysis.firstkids — Method.

firstkids(samples::Vector{<:NamedTuple}; skipethanol=true)
firstkids(samples::Vector{<:AbstractString}; skipethanol=true)

From a list of sample ids, identify the earliest sample for each child. Samples may be sample IDs that can be parsed by resolve_sampleID, or NamedTuples containing :subject and :timepoint fields.

If skipethanol is true, sample names containing _#E_ will be skipped.

Returns a vector of indicies that can be used to slice the original vector.

source

ECHOAnalysis.formulafeeding — Method.

formulafeeding(row::DataFrameRow)

Checks a wide-form metadata row for formula feeding information. Returns true if any of the following are true:

:milkFeedingMethods contains "Formula" (or "formula")

:exclusivelyFormulaFed is "Yes" (or "yes")

Both :exclusivelyNursed and :exclusiveFormulaFed are "No" (or "no")

:amountFormulaPerFeed is > 0:

Either :amountFormulaPerFeed or :formulaTypicalType have (non-missing) values

Otherwise returns false.

source

ECHOAnalysis.getfocusmetadata — Method.

getfocusmetadata(longfilepath, samples; focus=metadata_focus_headers)

Get a wide-form metadata table with a subset of headers for a subset of subject/sample IDs

source

ECHOAnalysis.getfocusmetadata — Method.

getfocusmetadata(df, samples; focus=metadata_focus_headers)

Get a wide-form metadata table with a subset of headers for a subset of subject/sample IDs

source

ECHOAnalysis.getmetadata — Method.

Get the metadata values for a given set of subjects and timepoints

source

ECHOAnalysis.getmetadatum — Function.

Get the metadata value for a given subject and timepoint

source

ECHOAnalysis.metacolor — Method.

metacolor(metadata::AbstractVector, colorlevels::AbstractVector; missing_color=colorant"gray")

Convert metadata into colors

source

ECHOAnalysis.numberify — Method.

numberify(x)
numberify(v::AbstractArray)

Try to convert something into a number or array of numbers using the heuristics:

if x is already a number or is missing, return x

if v is already an array of numbers, return v

if x is a String:

if x contains a . or e, parse as a Float64

Otherwise parse as an Int

if numberified vector contains any Floats, return a vector of missing and Float64

if x is anything other than a String or a number, return missing

source

ECHOAnalysis.resolve_letter_timepoint — Method.

resolve_letter_timepoint(sid::AbstractString)
resolve_letter_timepoint(sid::Missing)

Deals with sampleIDs in which the timepoint is given as a letter. Eg. 40b indicates timepoint 2 from subject ID 40. If no letter is provided, timepoint is assumed to be 1.

source

ECHOAnalysis.resolve_sampleID — Method.

resolve_sampleID(sid::Union{AbstractString,Symbol})

Parse a sample name and return its components as a named tuple:

Example: For the sample C0040_3F_1A:

C0040 means ""Child, subject ID 40"

3F means "timepoint 3", Fecal (genotech). May also be E for Ethanol.

1A means "replicate 1, aliquot A"

This function returns the NamedTuple (sample="C0040_3F_1A", subject=40, timepoint=3)

source

ECHOAnalysis.samplelessthan — Method.

samplelessthan(x::T, y::T) where T <: NamedTuple
samplelessthan(x::T, y::T) where T <: AbstractString

Function to use for sorting NamedTuples from sample IDs. String versions of sampleIDs will be parsed with resolve_sampleID

source

ECHOAnalysis.uniquesamples — Function.

uniquesamples(samples::AbstractVector{<:NamedTuple}, identifiers::Vector{Symbol}=[:subject,:timepoint];
                    skipethanol=true, samplefilter=x->true)

source

Index

ECHOAnalysis.add_batch_info

ECHOAnalysis.breastfeeding

ECHOAnalysis.firstkids

ECHOAnalysis.formulafeeding

ECHOAnalysis.getfocusmetadata

ECHOAnalysis.getfocusmetadata

ECHOAnalysis.getmetadata

ECHOAnalysis.getmetadatum

ECHOAnalysis.merge_tables

ECHOAnalysis.metacolor

ECHOAnalysis.numberify

ECHOAnalysis.resolve_letter_timepoint

ECHOAnalysis.resolve_sampleID

ECHOAnalysis.samplelessthan

ECHOAnalysis.uniquesamples