Subject and Sample Metadata

Subject and Sample Metadata

Functions

Index

Functions

ECHOAnalysis.breastfeeding — Method

breastfeeding(row::DataFrameRow)

Checks a wide-form metadata row for breastfeeding information. Possible return values are

:breastfed: at least 3 months exlussively breastfed, or where at least 80% of feedings are from breast

:formulafed: no breastmilk after 2 weeks

:mixed: all others

missing: not enough information

ECHOAnalysis.numberify — Method

numberify(x)
numberify(v::AbstractArray)

Try to convert something into a number or array of numbers using the heuristics:

if x is already a number or is missing, return x

if x is a String:

if x contains a . or e, parse as a Float64

Otherwise parse as an Int

if x is anything other than a String or a Number, return missing

if v is already an array of numbers (and missing), return v

if numberified vector contains any Floats, return a vector of missing and Float64

ECHOAnalysis.uniquetimepoints — Method

uniquetimepoints(samples::AbstractVector{<:StoolSample};
                    skipethanol=true,
                    samplefilter=x->true,
                    sortfirst=true,
                    takefirst=true)

Identifies unique timepoints (that is, subject=>timepoint pairs) from a vector of [AbstractTimepoint]@ref.

Example:

Given the following array of 4 samples:

s = stoolsample(["C0001_1F_1A", "C0001_1F_2A", "C0001_2F_1A", "C0002_1F_1A"])

The first three samples are from the same subject, the first two of which are from the same timepoint (they're replicates).

By default, this function excludes duplicate subject-timepoints, but keeps multiple timepoints for each individual subject. If a single timepoint from each subject is desired, use takefirst=true. In this later case, leave sortfirst=true (the default) to be sure to take the earliest timepoint for each subject.

Other parameters

skipethanol=true: exlude samples that match the pattern _\dE_, that is ethanol (as opposed to genotek) samples.

samplefilter=x->true: a function to select samples to include. By default, all samples are included. Use samplefilter=iskid to include only child samples for example.

sortfirst=true: if true, sorts the vector at the beginning

takefirst=false: if true, only takes 1 timepoint for each subject. Use with sortfirst to get only the first timepoint.

ECHOAnalysis.widemetadata — Method

widemetadata(longdf::AbstractDataFrame, samples::Vector{<:StoolSample};
                    metadata::Set=Set(unique(longdf.metadatum)),
                    parents::Set=Set(unique(longdf.parent_table)))
widemetadata(longdf::AbstractDataFrame, samples::Vector{<:AbstractString}; kwargs...)
widemetadata(db::SQLite.DB, tablename, samples; kwargs...)

Convert data in a long-form metadata table into wide-form, where each sample is given a single row and each metadatum a single column.

Timepoint-specific metadata (specifically, metadata that comes from a parent table that has a timepoint field) is only associated with samples from that timepoint, while others (eg. childGender) are associated with every sample for a given subject.

Optional key word args: - Pass vector of Parent Tables (parents) to include - Pass vector of metadatum fields (metadata) to include

Index

ECHOAnalysis.AbstractTimepoint

ECHOAnalysis.StoolSample

ECHOAnalysis.add_functional_profiles

ECHOAnalysis.add_taxonomic_profiles

ECHOAnalysis.breastfeeding

ECHOAnalysis.iskid

ECHOAnalysis.ismom

ECHOAnalysis.isstoolsample

ECHOAnalysis.numberify

ECHOAnalysis.replicateid

ECHOAnalysis.resolve_letter_timepoint

ECHOAnalysis.sampleid

ECHOAnalysis.sampletable

ECHOAnalysis.sampletype

ECHOAnalysis.sqlprofile

ECHOAnalysis.stoolsample

ECHOAnalysis.subject

ECHOAnalysis.timepoint

ECHOAnalysis.uniquetimepoints

ECHOAnalysis.widemetadata