Dealing with Metadata

Subject and Sample Metadata

Subject and Sample Metadata

Functions

Index

Functions

ECHOAnalysis.breastfeeding — Method.

breastfeeding(row::DataFrameRow)

Checks a wide-form metadata row for breastfeeding information. Returns true if any of the following are true:

:milkFeedingMethods contains "breast"

:exclusivelyNursed is "Yes" (or "yes")

Both :exclusivelyNursed and :exclusiveFormulaFed are "No" (or "no")

Any of the following have values > 0:

:typicalNumberOfFeedsFromBreast

:typicalNumberOfEpressedMilkFeeds

:lengthExclusivelyNursedMonths

:noLongerFeedBreastmilkAge

Otherwise returns false.

source

ECHOAnalysis.formulafeeding — Method.

formulafeeding(row::DataFrameRow)

Checks a wide-form metadata row for formula feeding information. Returns true if any of the following are true:

:milkFeedingMethods contains "Formula" (or "formula")

:exclusivelyFormulaFed is "Yes" (or "yes")

Both :exclusivelyNursed and :exclusiveFormulaFed are "No" (or "no")

:amountFormulaPerFeed is > 0:

Either :amountFormulaPerFeed or :formulaTypicalType have (non-missing) values

Otherwise returns false.

source

ECHOAnalysis.getfocusmetadata — Method.

getfocusmetadata(longfilepath, samples; focus=metadata_focus_headers)

Get a wide-form metadata table with a subset of headers for a subset of subject/sample IDs

source

ECHOAnalysis.getfocusmetadata — Method.

getfocusmetadata(df, samples; focus=metadata_focus_headers)

Get a wide-form metadata table with a subset of headers for a subset of subject/sample IDs

source

ECHOAnalysis.getmetadata — Method.

Get the metadata values for a given set of subjects and timepoints

source

ECHOAnalysis.getmetadatum — Function.

Get the metadata value for a given subject and timepoint

source

ECHOAnalysis.metacolor — Method.

metacolor(metadata::AbstractVector, colorlevels::AbstractVector; missing_color=colorant"gray")

Convert metadata into colors

source

ECHOAnalysis.numberify — Method.

numberify(x)
numberify(v::AbstractArray)

Try to convert something into a number or array of numbers using the heuristics:

if x is already a number or is missing, return x

if v is already an array of numbers, return v

if x is a String:

if x contains a . or e, parse as a Float64

Otherwise parse as an Int

if numberified vector contains any Floats, return a vector of missing and Float64

if x is anything other than a String or a number, return missing

source

ECHOAnalysis.resolve_letter_timepoint — Method.

resolve_letter_timepoint(sid::AbstractString)
resolve_letter_timepoint(sid::Missing)

Deals with sampleIDs in which the timepoint is given as a letter. Eg. 40b indicates timepoint 2 from subject ID 40. If no letter is provided, timepoint is assumed to be 1.

source

ECHOAnalysis.resolve_sampleID — Method.

resolve_sampleID(sid::Union{AbstractString,Symbol})
resolve_sampleID(sids::Vector{<:Union{AbstractString,Symbol}})

Parse a sample name and return its components as a named tuple:

Example: For the sample C0040_3F_1A:

C0040 means ""Child, subject ID 40"

3F means "timepoint 3", Fecal (genotech). May also be E for Ethanol.

1A means "replicate 1, aliquot A"

Hyphens in sample names are converted to underscores (so "C00011F1A" == "C0001-1F-1A")

Samples may also end with _S\d+ (well #s from sequencing) which are ignored

Samples that don't match the expected pattern return (sample=sid, subject=nothing, timepoint=nothing)

This function returns the NamedTuple (sample="C0040_3F_1A", subject=40, timepoint=3)

source

ECHOAnalysis.samplelessthan — Method.

samplelessthan(x::T, y::T) where T <: NamedTuple
samplelessthan(x::T, y::T) where T <: Union{AbstractString, Symbol}

Function to use for sorting NamedTuples from sample IDs. String versions of sampleIDs will be parsed with resolve_sampleID

source

ECHOAnalysis.uniquesamples — Method.

uniquesamples(samples::AbstractVector{<:NamedTuple}, identifiers::Vector{Symbol}=[:subject,:timepoint];
                    skipethanol=true, samplefilter=x->true)

Identifies unique samples from a vector of sample tuples. Tuples must have a sample field, and additional fields specified in identifiers. Multiple items that have the same values among the identifiers will be excluded.

Example:

Given the following array of 4 samples:

s = [(sample="C0001_1F_1A", subject=1, timepoint=1),
     (sample="C0001_1F_2A", subject=1, timepoint=1),
     (sample="C0001_2F_1A", subject=1, timepoint=2)
     (sample="C0002_1F_1A", subject=2, timepoint=1)];

By default, identifiers = [:subject, :timepoint], meaning only the second item would be excluded (since both it and the first item are subject 1, timepoint 1).

A single timepoint for each subject can be acquired by passing identifiers=[:subject].

uniquesamples(s)

uniquesamples(s, identifiers=[:subject])

Other parameters

skipethanol=true: exlude samples that match the pattern _\dE_, that is ethanol (as opposed to genotek) samples.

samplefilter=x->true: a function to select samples to include. By default, all samples are included. Use samplefilter=x->startswith(x, "C") to include only child samples for example.

takefirst=true: if true, sorts the samples using [samplelessthan]@ref

source

Index

ECHOAnalysis.add_batch_info

ECHOAnalysis.breastfeeding

ECHOAnalysis.formulafeeding

ECHOAnalysis.getfocusmetadata

ECHOAnalysis.getfocusmetadata

ECHOAnalysis.getmetadata

ECHOAnalysis.getmetadatum

ECHOAnalysis.merge_tables

ECHOAnalysis.metacolor

ECHOAnalysis.numberify

ECHOAnalysis.resolve_letter_timepoint

ECHOAnalysis.resolve_sampleID

ECHOAnalysis.samplelessthan

ECHOAnalysis.uniquesamples